Merge samba.org:/scratch/anton/linux-2.5 into samba.org:/scratch/anton/for-alan
2003/01/15 22:49:30-08:00 randy.dunlap
[PATCH] update LOG BUF SIZE config.

The current LOG_BUF size is a bit confusing the first
time that "make oldconfig" is used.  It's difficult to
select anything other than the default value.

Also, you (Linus) expressed a desire to have this
configurable only if DEBUG_KERNEL or "kernel hacking"
was enabled, so I've changed it to accomplish that.

This patch also uses Kconfig in a way that Roman intended
since a patch in 2.5.52 which enables default values if
a prompt is not enabled, but lets values be chosen when
the prompt is enabled.  You also asked for this in setting
this config option.

2003/01/15 22:27:24-08:00 torvalds
Merge bk://ppc.bkbits.net/for-linus-ppc
into home.transmeta.com:/home/torvalds/v2.5/linux

2003/01/16 15:15:10+11:00 paulus
PPC32: Use a per-cpu variable for prof_counter and prof_multiplier.

2003/01/16 15:10:04+11:00 paulus
PPC32: Better support for PPC 4xx debug facilities.

This provides for separate global and per-thread debug control
register value(s), which are switched as appropriate.  This allows
us to use both an external JTAG debugger for debugging the kernel
as well as using gdb to debug user programs.

2003/01/16 14:52:58+11:00 paulus
PPC32: Page-align the data section of the boot wrapper.

This is needed for Open Firmware on older powermacs to be able to
load the wrapper.  Without this OF gives a "CLAIM failed" error.

2003/01/16 14:47:30+11:00 paulus
PPC32: Add support for PPC 4xx on-chip devices using the generic
device model.

2003/01/16 11:52:03+11:00 paulus
Merge bk://68.0.152.218/linux-2.5-misc
into samba.org:/home/paulus/kernel/for-linus-ppc

2003/01/15 16:48:27-07:00 trini
PPC32: Change the MontaVista copyright / GPL boilerplate to
a condensed version.

BKrev: 3e267b2btwSTxAFoON33xre5Z3gy5Q