Merge kanga.twiddle.net:/home/rth/linux/linus-2.5 into kanga.twiddle.net:/home/rth/linux/axp-2.5
2003/01/16 11:46:25-08:00 mbligh
[PATCH] (3/3) NUMA rebalancer

Patch from Erich Focht

This adds a hook to rebalance globally across nodes every NODE_BALANCE_RATE
iterations of the rebalancer. This allows us to easily tune on an architecture
specific basis how often we wish to rebalance - machines with higher NUMA
ratios (more expensive off-node access) will want to do this less often.
It's currently set to 100 for NUMA-Q and 10 for other machines. If the
imbalance between nodes is > 125%, we'll rebalance them. The hook for this
is added to the NUMA definition of cpus_to_balance, so again, no impact
on non-NUMA machines.

2003/01/16 11:46:10-08:00 mbligh
[PATCH] (2/3) Initial load balancing

Patch from Michael Hohnbaum

This adds a hook, sched_balance_exec(), to the exec code, to make it
place the exec'ed task on the least loaded queue. We have less state
to move at exec time than fork time, so this is the cheapest point
to cross-node migrate. Experience in Dynix/PTX and testing on Linux
has confirmed that this is the cheapest time to move tasks between nodes.

It also macro-wraps changes to nr_running, to allow us to keep track of
per-node nr_running as well. Again, no impact on non-NUMA machines.

2003/01/16 11:45:56-08:00 mbligh
[PATCH] (1/3) Minimal NUMA scheduler

Patch from Martin J. Bligh

This adds a small hook to the find_busiest_queue routine to allow us to
specify a mask of which CPUs to search over. In the NUMA case, it will
only balance inside the node (much cheaper to search, and stops tasks
from bouncing across nodes, which is very costly). The cpus_to_balance
routine is conditionally defined to ensure no impact to non-NUMA machines.

This is a tiny NUMA scheduler, but it needs the assistance of the second
and third patches in order to spread tasks across nodes.

2003/01/16 11:38:42-08:00 hch
[PATCH] remove MOD_IN_USE

Another left-over from ancient module code, it was supposed to return
non-zero if the module has a use count, but currently it always
evaluates to 0.

There are a few users of different types:
(1)  ioctl that perform a while(MOD_IN_USE) MOD_DEC_USE_COUNT loop.
     Just rip them out, we now have forced module unloading.
(2)  printk's that moan if the use-count in not zero in the exitfunc.
     Just rip them out, this can't happen.
(3)  if(MOD_IN_USE) MOD_DEC_USE_COUNT constructs in ->close of a few
     serial drivers.  Just remove the conditional, we did a
     MOD_INC_USE_COUNT in ->open.
(4)  This one is interesting:  drivers/sbus/char/display7seg.c uses
     the module use count to track openers.  Replace this with an
     atomic_t.

In addition remove tons of stale comments in network driver that aren't
understandable for anyone who doesn't know ancient Linux module semantics.

2003/01/16 11:36:37-08:00 ak
[PATCH] x86_64 update

x86-64 updates for 2.5.58. Changes only x86-64 specific files.

 - Rewrote module allocation. Lots of bugs fixed. Module loading
   should work now again.
 - Kconfig help fixes from Randy Dunlap
 - Makefile cleanups from Pavel Machek and Sam Ravnborg
 - Assembly cleanups from Pavel
 - defconfig update
 - Better strlen_user/strnlen_user
 - Merge with i386: new ptrace commands, 32bit vsyscall signal trampolines
		new deactivate_mm, add asm/bug.h
 - Make sure initramfs is freed after booting (thanks to Kai for the hint)
 - User per cpu data for profile counters (Ravikiran Thirumalai)
 - 32bit compat_* updates from Stephen Rothwell
 - Fix race in context switch. The exception handler for bogus segment
   loads in __switch_to needs to keep interrupts disabled, otherwise an
   interrupt can deadlock on scheduler locks.  Also make sure they don't
   printk or set oops_in_progress during printk because printk does a
   wake_up too.
 - Disable 64bit GS base changes for processes.  I cannot get it to work
   reliably.
 - Clear IOPL on kernel entry

2003/01/16 10:23:42-08:00 torvalds
Merge http://linux-voyager.bkbits.net/eisa-sysfs-2.5
into home.transmeta.com:/home/torvalds/v2.5/linux

2003/01/16 09:28:13-08:00 hch
[PATCH] remove GET_USE_COUNT

This is a left-over from the old modules code, Rusty stubbed it out
to always return 0.  Three scsi pcmcia driver check it for beeing non-NULL,
trying to work around their unload races.  I've added #warnings there
and stubbed out the GET_USE_COUNT so we can remove it from the core.

2003/01/16 09:28:04-08:00 hch
[PATCH] fix signed/unsigned issue in SGI partitioning code

The Linux code for SGI partitions uses an int instead of an unsigned int
in the ondisk structure in two places, which breaks > TB partitions.

While porting the code over from an internal 2.4-based tree I've also
switched it to use the explicit uXX/sXX types everywhere and moved the
struct defintions above sgi_partition().

2003/01/16 09:41:26-06:00 mzyngier
EISA sysfs AIP update

Without it, unloading a module
leads to some unpleasant oops...

2003/01/16 09:21:02-06:00 mzyngier
EISA sysfs updates to 3c509 and 3c59x drivers

2003/01/16 09:19:30-06:00 mzyngier
EISA naming database

Please note that the naming DB is now completely optional. If there is
no eisa.ids in the drivers/eisa/ directory, build will behave as if
CONFIG_EISA_NAMES is disabled. So this patch can be left out if there
is any objection.

2003/01/16 09:18:00-06:00 mzyngier
sysfs EISA support

Base patch adding sysfs support for the EISA bus

BKrev: 3e275902Lx9i2vuGvUDd3f6N3bVF-A