[PATCH] Restore LSM hook calls to sendfile
Patch from "Stephen D. Smalley" <sds@epoch.ncsc.mil>

This patch restores the LSM hook calls in sendfile to 2.5.59.  The hook was
previously added as of 2.5.29 but the hook calls in sendfile were
subsequently lost as a result of the sendfile rewrite as of 2.5.30.

BKrev: 3e3ff1a3eDhHHdMI2zu5g8Uo1nkCcA