[PATCH] revert extra sendfile security hook patch
hm.  It seems that I sent this patch twice.

After resyncing with your tree I go through and try to reapply all the sent
patches, throwing out the ones which get a lot of rejects.  Just to make sure
that everything got through OK.

But it appears that that particular patch happily applied on top of itself,
so I assumed it was not applied...

BKrev: 3e42fc2c-tS8lonEYEXy64fCFeKViQ