[SCSI] Migrate sim710 to 53c700 chip driver
This should add synchronous support and Tagged Command Queueing.
At the moment, it cuts down on the number of command line options,
but we can add those back in later.

This patch also migrates the driver to the new device model for
both MCA and EISA.

BKrev: 3e467cd7dY8p8nEe_CfjWrywUuKWaA