Merge http://linux-isdn.bkbits.net/linux-2.5.make into home.transmeta.com:/home/torvalds/v2.5/linux
2003/02/08 14:33:21-06:00 kai
kbuild: Add a bug trap for people playing with SUBDIRS too much

If SUBDIRS is set manually on the command line, the contents of
.tmp_versions are not deleted before descending and can accumulate
stale entries. Print a warning if that case is detected, but deal with
it gracefully.

2003/02/08 14:30:37-06:00 kai
kbuild: Modversions fix

We're still using the old genksyms binary, that's why we have to
postprocess the output to convert it into a linker script - that
postprocessing got confused by "__verify_write".

Kick out the grep, do it all and correctly within sed.

Bug reported by Thomas Molina.

2003/02/08 14:00:46-06:00 kai
kbuild: Warn on obsolete export-objs use

Setting export-objs is not necessary anymore, so warn on encountering it
to prevent it from creeping back in ;)

Also, make the error when we find someone still using O_TARGET non-fatal,
so that people sharing stuff between 2.4 and 2.5 don't have more hassle
than necessary.

2003/02/08 13:59:13-06:00 kai
kbuild: Handle external SUBDIRS with modversions

We need to collect a list of all modules during the recursive build. I used
a "touch .tmp_versions/<path/to/module.ko>" to do so, which however doesn't
work so well, when path/to isn't inside the kernel tree.

The best way to build external modules is currently using kbuild by saying
"make SUBDIRS=/some/external/dir modules", which was thus broken. While this
way is not all that optimal and I hope to come up with something better
before 2.6, it works and should keep working, so this patch fixes the usage
above.

Instead of touching files with the entire path added, we just create a
<module>.mod file in $(MODVERDIR) now, and save the path to the module.ko
in it. Since module names are unique, a flat hierarchy is actually fine here.

BKrev: 3e4712beqP259xaaxxF_rKe931l98Q