[PATCH] direct-io: allow reading of the part-filled EOF block
driect-io will currently return EINVAL when the application tries to read the
final bit of the file at EOF.  (assuming the file's length is not a multiple
of the filesystem blocksize).

The 2.4 kernelwill reurn 0 (it won't read it at all).

This patch changes the 2.5 kernel to allow that block to be read.

BKrev: 3e4ea1dcQIync8OvXnum6MMxFpkb9A