[PATCH] Add printk levels to mtrr, also clarify
The big thing here is actually turning
'Your bios sucks'
into
'Your bios sucks, but we fixed it and life is good, don't panic'

Which with a vendor hat on is important.  Its currently hard to tell
whether some Linux errors are things to worry about or merely spanking
the manufacturer.  I'd urge people writing error messages to think about
that btw - does it tell the user if the problem is fixed ?

BKrev: 3e4ea348F2EBEqw3vJ4IncGwXqVuCA