[PATCH] pcmcia: add device_class pcmcia_socket, update devices & drivers
A new device_class "pcmcia_socket_class" is introduced for PCMCIA and
CardBus sockets.  All socket drivers I could find are updated so that
they register a driver, and -if necessary- the "platform"/legacy device.
This will allow for a cleanup of pcmcia_{un}register_socket() /
{un}register_ss_entry() as well as reflect the parent for pcmcia_bus
devices.

Russell King allowed me to break sa1100_generic pcmcia support for the
time being - so drop that part of the patch for the moment.

BKrev: 3e4ea6446P-5b1LylR7AQyf8dzofkg