[PATCH] export add_to_page_cache() and __pagevec_lru_add to
CIFS is using these.

Given that the readpages() address_space op is supposed to add the pages to
pagecache, it makes sense to make these functions available to modules.

I can't say that I put a lot of though into the readpages API.  It was
designed as just enough functionality to be able to stuff a bunch of
readahead pages into a single BIO.  The only reason I made it an a_op at all
was because we have toi enter the fs to pick up the ->get_block callback's
address.

But a couple of filesystems seem to be making use of it now.  Reiser4 will
need access at the do_page_cache_readahead() level too.

BKrev: 3e53151a-YaqKCDdKwHQkfBn3w4eww