[PATCH] use local RODATA setup for m68knommu linker script
This patch removes the use of the common RODATA define in the m68knommu
architecture. It cannot be used the same way for the m68knommu target.
For starters just inserting it here is syntactically wrong. All the read
only parts are grouped into a single "text" segment, and this is the root
cause of the problem. So for the m68knommu arch it makes sense to not
use the generic RODATA setup, but to list them locally.

BKrev: 3e531a6emKisBlssyk_vR2_1Uiac5g