[PATCH] create an architecture specific flat header for m68knommu
Add a per-architecture flat.h flat format executable header. The idea is
to support the different options required in a clean way. For starters
we need to be able to configure the initial stack layout, and this
differs for different CPU types.

This first patch adds the header for the m68knommu architecture.

BKrev: 3e531a8bJYGvKtSJLf2VmBgwXU4XXg