Merge davem@nuts.ninka.net:/home/davem/src/BK/net-2.5 into kernel.bkbits.net:/home/davem/net-2.5
2003/02/24 03:18:46-08:00 davem
[IPV{4,6}]: Make icmp_socket per-cpu and simplify locking.

2003/02/23 23:13:18-08:00 davem
[IPV6]: ipv6_count_addresses is static.

2003/02/23 23:11:30-08:00 davem
[IPV6]: Cure typo in ipv6_addr_prefix.

2003/02/23 22:46:42-08:00 yoshfuji
[IPV6]: Privacy Extensions for Stateless Address Autoconfiguration.

2003/02/23 22:42:02-08:00 yoshfuji
[IPV6]: Export ip6_route_me_harder for netfilter and add ipv6_addr_prefix.

2003/02/23 22:36:41-08:00 chas
[ATM]: prevent compiler warning when compiling w/o bridging

2003/02/23 22:34:55-08:00 chas
[ATM]: empty tx queue in lec when flush complete
the lane spec says the following about path switching:

10.1.2.7 Transmitting Held Frames

Once the LE_FLUSH_RESPONSE is received, the LE Client MUST
transmit any held  data frames on the new data path before
transmitting any further frames on the new path.

10.1.2.8 Switching Over Paths Without Flush

Regardless of the provisions of the rest of Section 10.1, when
switching from the old path to a new path, if an LE Client has
not transmitted a data frame to a particular LAN Destination via
the old path for a period of time greater than or equal to the
C22 Path Switching Delay, then it MAY start using the new path
without employing the Flush protocol.

the lane client in linux-atm does do this but it doesnt send the
held frames until a frame is transmitted on the vcc after the flush
has completed.  often this doesnt happen until you get an ip retransmit.
the following patch sends the held frames after the flush completes
(or the switching delay expires)

2003/02/23 22:33:05-08:00 chas
[ATM]: Remove cli from lec.c

2003/02/23 22:14:23-08:00 jmorris
[NETLINK]: Un-duplicate rcv wakeup logic.

2003/02/23 01:50:01-08:00 davem
[BLUETOOTH]: net/bluetooth/bnep/sock.c needs linux/init.h

2003/02/22 22:20:14-08:00 davem
[NETLINK]: Remove buggy and useless rcv queue wakeup tests.

2003/02/22 21:57:30-08:00 jmorris
[NET]: Add myself as co-maintainer.

2003/02/22 16:15:45-08:00 hch
[NET]: Remove __NO_VERSION__ from networking code.

2003/02/22 16:12:51-08:00 ak
[NET]: Make skbuff.h -W clean, skb_headlen should return unsigned quantity.

BKrev: 3e59ff9789dar3rTKRjcP_Tn9qhTxQ