[PATCH] Early ioremap support for ia32
Patch from Patricia Gaughen <gone@us.ibm.com>,
		Dave Hansen <haveblue@us.ibm.com>

It provides a very early sort of kmap-by-hand.

The patch is used by the x440 discontigmem to map the srat tables into low
memory so that the memory can be setup.  This remap function is used very
early in the boot process...  at the start of setup_arch().

This functionality is only available to Summit and NUMAQ.  It will work on
other platforms, but they do not need it.

BKrev: 3e5ba26cMtnQGGvcdif44ZwUqtIQSg