[PATCH] Modules code tidy up
Now that we search for 11 different sections by name, the if/else was
getting unwieldy.  Also, handle_section just does relocs, so it's a bit
of a misnomer, and it's best simply moved into the main code.

This open-codes handle_section, which simply does relocations now.

Also adds "find_sec" and uses it to find the various sections.

BKrev: 3e5c25d3F2pPhbCfcQrwsVSbre8wVA