[PATCH] replace its with it's where appropriate.
This patch replaces its (possessive of it) with it's (it is)
in the following cases where "it is" is meant.

its a   -> it's a
its an  -> it's an
its not -> it's not

except for the files
   arch/cris/boot/rescue/head.S
   arch/cris/kernel/kgdb.c
where the substitution is "its not" -> "it is not"
to avoid possible problems with single quotes in assembly comments.

BKrev: 3e5c2b49xZMuQJJjg4bgwSZ40mwymw