[PATCH] fix SMP lockup in eepro100 with ethtool on unused
Patch from Jason Lunz <lunz@falooley.org>

When support for the GSET and SSET ethtool ioctls was added to eepro100.c in
2.4.20, the tx lock was overloaded to serialize their use.  Unfortunately,
this lock is only initialized in dev->open(), causing ethtool to deadlock the
machine when used on an unconfigured eepro100 interface.

The fix is to initialize the spinlock at probe time.

BKrev: 3e6ab4delwxArUdaJiNmPd_eRHe0jA